pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "investment_agent"
version = "0.1.0"
description = "Property Investment Analysis App with AI Agents"
readme = "README.md"
authors = [
    {name = "Property Investment Team"},
]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # Base dependencies
    "fastapi>=0.108.0", # Updated to support anyio 4.5+
    "uvicorn==0.23.2",
    "pydantic>=2.0.0", # Added from requirements.txt
    "python-dotenv==1.0.0",
    "sqlalchemy==2.0.23",
    "alembic==1.12.1",
    # Frontend related
    "streamlit==1.28.1",
    "plotly==5.18.0",
    "pandas==2.1.2",
    "numpy==1.26.1",
    # AI Agent dependencies
    "langchain>=0.1.0", # Updated to support anyio 4.5+
    "langchain-core>=0.1.0", # Added from requirements.txt
    "langchain-community>=0.0.11", # Updated for compatibility with newer langchain
    "langchain-experimental>=0.0.9", # Added from requirements.txt
    "langgraph>=0.0.19", # Added from requirements.txt
    "openai>=1.6.0", # Updated version
    "anthropic>=0.8.0", # Updated to support anyio 4.5+
    "chromadb==0.4.18",
    "chroma-hnswlib", # Added from requirements.txt
    "faiss-cpu", # Changed from faiss-gpu to faiss-cpu for Python 3.11 compatibility
    "sentence-transformers>=2.2.2", # Added from requirements.txt
    "beautifulsoup4==4.12.2",
    "requests==2.31.0",
    "selenium==4.14.0",
    "scrapy==2.11.0",
    "pypdf==3.17.0",
    "langsmith>=0.0.43", # Added from requirements.txt
    "openai-agents==0.0.13", # OpenAI Agents SDK
    "pydantic-settings>=2.5.2", # Required for openai-agents
    "typing-extensions>=4.12.2", # Required for openai-agents
    "azure-identity>=1.15.0", # For Azure OpenAI authentication
    "logfire>=3.14.1",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.8",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "black==23.10.1",
    "isort==5.12.0",
    "mypy==1.6.1",
    "ruff==0.1.3",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/investment_agent"
"Bug Tracker" = "https://github.com/yourusername/investment_agent/issues"

[project.scripts]
investment_agent = "investment_agent.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/investment_agent"]

[tool.hatch.build.targets.sdist]
include = [
    "src/investment_agent",
    "README.md",
    "LICENSE",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "B"]
ignore = []

[tool.uv]
# Configuration options for uv package manager
requires-python = ">=3.11"
editable-mode = "development"
